Fix multipart grouping.
This worked last time because I ran the wrong binary and things looked
good and noticed that when the auth failed and rerunning didn't
finish.  The new binary finished fine.

However, due to port naming (multipart parts have parts and naming is
hard), I grouped incorrectly, so I was only attempting to upload a
single part at a time.