Use AWS Lambda interface directly instead of via SQS
There's no useful way to tell whether invocation completed or not.
Lambda has destination SQS queues for success and failure, but they
are not used when the function is invoked via SQS.  I find
this... weird.

They are, however used when the function is explicitly invoked
asynchronously, which is what I'm doing here.

I did a pass with synchronous invocation of the lambda functions.  It
was kind of nice in that I got to observe the state, but it's net
worse because if the request times out or otherwise fails, I end up
with a *partial* copy.  So there's no real way around a good solution
requiring local state.