Updated to the latest gopro-plus.
Part of doing this allowed me to greatly simplify extractSources, for
which I also wrote a unit test since it's kind of important to the
whole backup plan.

It'll also be obvious what I need to do for adding stuff from the
"files" section because I might want that, too even though it's kind
of weird.