Optimize processSQS with more concurrency and better batching
I ran a test in which I simulated a complete copy of all my artifacts
from GoPro to S3.  For 2,486 uploaded media items totaling a bit over
1 TiB, I copied 6,116 artifacts totaling around 1.28 TiB.  The
overhead difference was around 256.8 GiB.  The simulation didn't
actually *copy* files, but merely ran through all the other parts of
the system, including asking GoPro Plus for signed URLs and issuing
HEAD requests to collect the details.

I found that the slowest part of my processing was actually in
gathering the responses via SQS.  You can only ask for 10 at a time,
so my old mechanism was:

1. Ask for up to 10.
2. For each update:
   a. Update state in local DB.
3. Ask Amazon to delete the 10 I got.

I didn't measure very specifically, but the delete seemed particularly
slow.  In any case, it seems issuing multiple queue requests in
parallel is a best practice here, so now the mechanism is
approximately this:

1. Compute a number of workers (N) appropriate for the outstanding
   file count.
2. Spawn N workers asking for queue results.  (0th worker polls, the
   rest only get items that are ready).
3. Update local state for all results in a single DB call.
4. Re-chunk results into their maximum batch size and issue deletes
   for all found results concurrently.

With a limit of 10 concurrent workers, this approximately allows 100
messagse to be processed per cycle instead of 10, but the felt delays
should be similar.