Allow pulling complete media rows from the DB.
This also provides access to the file fields through a lens over the
JSON.  That required a lib update because I needed to be able to
remarshal that lens.

Also, I no longer actually *remove* the URLs from the downloaded
metadata as it makes aeson unhappy when trying to parse them.
Instead, I just set the URLs to "".  A little lame, but it maintains
structure while throwing away large, unusable bits of data.