Process local backups in query order.
Converting both to sets to do the difference gives me alphabetical
order by medium ID which is probably better from a randomness
perspective, but I can do that in the query if I want it.  I wrote the
query to sort by oldest-first, so I might as well process the stuff in
that order.