Stop using gpmf-only export stream for metadata.
This is ideal, and fast, but, we currently expect to get all the data
from a contiguous stream and we're getting non-concatengated gpmf
streams currently.  More refactoring is required to allow for multiple
streams.  The extraction from the streams is straightforward, but the
thing that names the bits is missing, so a bit more work than I want
to do is required.