store s3 metas earlier in the sync
There's no cheap way to determine what's on S3, so the meta thing is a
proxy for blobs that need to be uploaded.

Having other potentially slow tasks between these can help prevent us
from moving stuff.

This doesn't solve the problem, but S3 is mostly an optimization for
some workflows, so less likely is an improvement.